django-admin/pyproject.toml
[tool.poetry]
name = "django-admin"
version = "0.1.0"
description = ""
authors = ["Ruslan Filichkin <ruslan.filichkin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "4.0.4"
asgiref = "3.5.2"
attrs = "21.4.0"
autopep8 = "1.6.0"
django-autocomplete-light = "3.9.4"
django-cors-headers = "3.13.0"
django-debug-toolbar = "3.5.0"
django-extensions = "3.2.0"
django-split-settings = "1.1.0"
djangorestframework = "3.13.1"
flake8 = "4.0.1"
gnureadline = "8.1.2"
gunicorn = "20.1.0"
iniconfig = "1.1.1"
mccabe = "0.6.1"
packaging = "21.3"
pluggy = "1.0.0"
psycopg2-binary = "2.9.3"
py = "1.11.0"
pycodestyle = "2.8.0"
pyflakes = "2.4.0"
pyparsing = "3.0.9"
pytest = "7.1.2"
python-dateutil = "2.8.2"
python-dotenv = "0.20.0"
pytz = "2022.1"
six = "1.16.0"
sqlparse = "0.4.2"
toml = "0.10.2"
tomli = "2.0.1"
uWSGI = "2.0.20"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
